# crates/notary/server/openapi.yaml
openapi: 3.0.0
info:
  title: Notary Server
  description: Notary server written in Rust to provide notarization service.
  version: 0.1.0-alpha.12
tags:
- name: General
- name: Notarization
paths:
  /healthcheck:
    get:
      tags:
      - General
      description: Healthcheck endpoint
      security:
        - {} # make security optional
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Ok response from server
          content:
            text/plain:
              schema:
                type: string
                example: Ok
        '401':
          description: API key is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 'Unauthorized request from prover: Invalid API key.'
  /info:
    get:
      tags:
      - General
      description: General information about the notary server
      security:
        - {} # make security optional
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Info response from server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
        '401':
          description: API key is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 'Unauthorized request from prover: Invalid API key.'
  /notarize:
    get:
      tags:
      - Notarization
      description: Start notarization for TCP or WebSocket client
      security:
        - {} # make security optional
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
      - in: header
        name: Connection
        description: The value should be 'Upgrade'
        schema:
          type: string
          enum:
          - Upgrade
        required: true
      - in: header
        name: Upgrade
        description: The value should be 'TCP'
        schema:
          type: string
          enum:
          - TCP
        required: true
      - in: query
        name: clientType
        description: Type of client (tcp or websocket)
        schema:
          type: string
          enum:
          - tcp
          - websocket
        required: true
      - in: query
        name: maxSentData
        description: Maximum data that can be sent by the prover (optional)
        schema:
          type: integer
        required: false
      - in: query
        name: maxRecvData
        description: Maximum data that can be received by the prover (optional)
        schema:
          type: integer
        required: false
      responses:
        '101':
          description: Switching protocol response
        '400':
          description: Headers or parameters provided by prover are invalid
          content:
            text/plain:
              schema:
                type: string
                example: 'Invalid request from prover: Upgrade header is not set for client'
        '401':
          description: API key is invalid
          content:
            text/plain:
              schema:
                type: string
                example: 'Unauthorized request from prover: Invalid API key.'
        '500':
          description: There was some internal error when processing
          content:
            text/plain:
              schema:
                type: string
                example: Something is wrong
        '503':
          description: Service is temporarily unavailable due to high load
          headers:
            Retry-After:
              description: Suggested retry time in seconds
              schema:
                type: integer
                example: 5
components:
  schemas:
    InfoResponse:
      type: object
      properties:
        version:
          description: Current version of notary server
          type: string
        publicKey:
          description: Public key of notary server for its notarization transcript signature
          type: string
        gitCommitHash:
          description: The git commit hash of source code that this notary server is running
          type: string
        quote:
          type: object
          properties:
            rawQuote:
              description: Hex bytes representing the signed-by-intel quote
              type: string
            mrsigner:
              description: Represents the public key of the enclave signer
              type: string
            mrenclave:
              description: The enclave image hash, including gramine and the notary server itself
              type: string
            error:
              description: Error that occurs when generating this quote
              type: string
      required:
      - version
      - publicKey
      - gitCommitHash
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Whitelisted API key if auth module is turned on and in whitelist mode
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JSON Web Token if auth module is turned on and in JWT mode